consolidate gummyroll state
